Detecting PS/2 devices

+++ src/utils/bitutils.rs
@@ -1,3 +1,3 @@
-pub fn is_nth_bit_set(number: u8, bit_index: u8) -> bool {
+pub fn is_nth_bit_set(number: u8, bit_index: usize) -> bool {
     number & (1 << bit_index) != 0
 }
